# PERF: DataFrame.insert